+++ regimes/in/org_item.go
@@ -1,0 +1,15 @@
+package in
+
+import (
+	"github.com/invopop/gobl/org"
+	"github.com/invopop/validation"
+)
+
+func validateOrgItem(it *org.Item) error {
+	return validation.ValidateStruct(it,
+		validation.Field(&it.Identities,
+			org.RequireIdentityType(IdentityTypeHSN),
+			validation.Skip,
+		),
+	)
+}